Fixes: qml file paths need file:/// on windows
Path format for qml use should be as url acording to
http://stackoverflow.com/questions/17610949/protocol-c-is-unknown-error-in-qml
While it is not nescesarry on linux, it doesn't seem to cause problmes either.
I am not sure this is the right way to solve this problem, actually I am pretty
sure there is a better more general Qt method, I just can't find it.